+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [0.26.0]
+
+### Added
+- Improved support for `aiclassify` with OpenAI models (you can now encode upto 40 choices).
+- Added a template for routing questions `:QuestionRouter` (to be used with `aiclassify`)
+
+### Changed
+- [BREAKING] Changed behaviour of `aitemplates(name::Symbol)` to look for the exact match on the template name, not just a partial match. This is a breaking change for the `aitemplates` function only. Motivation is that having multiple matches could have introduced subtle bugs when looking up valid placeholders for a template.
+
 ## [0.25.0]
 
 ### Added